Project Init and first build of ColorQuest
The Unity project was initiated and the first build of ColorQuest was added, but not in the right folder right away, because there might be serveral changes in terms of mechanics (aiming, moving, etc.; our aim algorithm might be even useless because of the controllers). We might have to begin this from scratch.